added react-router to app
